+++ Scribe/Scribe-Core/Changes.md
@@ -2,6 +2,9 @@
 
 ## v1.0.1 (in development)
 Supports **Flow v2.3**
+### New features
+- **Synagogue** now supports "copy-logging"
+  - Simply call `.copyTo(...)` in order to introduce logging end-points that also need to be notified of any records
 ### Other changes
 - Scala version updated to 2.13.12