+++ Scribe/Scribe-Core/Changes.md
@@ -3,6 +3,8 @@
 ## v1.0.1 (in development)
 Supports **Flow v2.3**
 ### New features
+- Added **ConsoleScribe** class that writes detailed issue data to the console, 
+  and possibly to an external log file as well
 - **Synagogue** now supports "copy-logging"
   - Simply call `.copyTo(...)` in order to introduce logging end-points that also need to be notified of any records
 ### Other changes