+++ Scribe/Scribe-Core/Changes.md
@@ -1,7 +1,7 @@
 # Utopia Scribe Core - list of changes
 
 ## v1.0.1 (in development)
-Supports **Flow v2.3**
+This update focuses on enabling you to print error data to console as well as to the default target (server database).
 ### New features
 - Added **ConsoleScribe** class that writes detailed issue data to the console, 
   and possibly to an external log file as well